requirements.txt
brotlipy==0.7.0
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1613246951236/work
chardet @ file:///tmp/build/80754af9/chardet_1607706775000/work
cryptography @ file:///tmp/build/80754af9/cryptography_1616767007030/work
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
olefile==0.46
Pillow @ file:///tmp/build/80754af9/pillow_1617385572585/work
pyaml @ file:///home/linux1/recipes/ci/pyaml_1610972911625/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
PyYAML==5.4.1
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
six @ file:///tmp/build/80754af9/six_1605205306277/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
